fix(architecture): extract roleNormalization to shared package
