+++ .jscpd-baseline.json
@@ -1,6 +1,6 @@
 {
-  "baselinePercentage": 0.15,
+  "baselinePercentage": 0.17,
   "maxAllowedPercentage": 0.25,
   "tolerance": 0.05,
-  "lastUpdated": "2026-07-25"
+  "lastUpdated": "2026-07-28"
 }
