+++ shared/src/utils/roleNormalization.ts
@@ -1,0 +1,69 @@
+import { Role } from '@esparex/contracts';
+
+/**
+ * Normalizes legacy role strings to canonical Role enum values.
+ * This provides a temporary compatibility layer during the migration phase.
+ * 
+ * @param role - The raw role string from database or JWT
+ * @returns The canonical Role value
+ */
+export const normalizeRole = (role: string | null | undefined): Role => {
+    if (!role) return Role.USER;
+
+    const normalized = role.toLowerCase().trim();
+
+    // 1. Super Admin variants
+    if (normalized === 'super_admin' || normalized === 'superadmin' || normalized === 'superAdmin') {
+        return Role.SUPER_ADMIN;
+    }
+
+    // 2. Business variants
+    if (normalized === 'business' || normalized === 'seller_pro') {
+        return Role.BUSINESS;
+    }
+
+    // 3. Moderator variants
+    if (
+        normalized === 'moderator' || 
+        normalized === 'support' || 
+        normalized === 'finance' || 
+        normalized === 'finance_manager' || 
+        normalized === 'user_manager' ||
+        normalized === 'content_moderator' ||
+        normalized === 'editor' ||
+        normalized === 'viewer'
+    ) {
+        return Role.MODERATOR;
+    }
+
+    // 4. Admin variants
+    if (normalized === 'admin') {
+        return Role.ADMIN;
+    }
+
+    // Default to USER
+    return Role.USER;
+};
+
+/**
+ * Checks if a role string represents a Super Admin account.
+ */
+export const isSuperAdminRole = (role: string | null | undefined): boolean => {
+    return normalizeRole(role) === Role.SUPER_ADMIN;
+};
+
+/**
+ * Checks if a role string represents any administrative role (Admin or Super Admin).
+ */
+export const isAdminRole = (role: string | null | undefined): boolean => {
+    const normalized = normalizeRole(role);
+    return normalized === Role.ADMIN || normalized === Role.SUPER_ADMIN;
+};
+
+/**
+ * Checks if a role string is legacy and needs normalization.
+ */
+export const isLegacyRole = (role: string): boolean => {
+    const canonicalValues = Object.values(Role) as string[];
+    return !canonicalValues.includes(role);
+};
